# gestion_canditature/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "endPoind ",
        "version": "1.0"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/formations/store": {
            "post": {
                "tags": [
                    "Formations"
                ],
                "summary": "Créer une nouvelle formation",
                "description": "Permet de créer une nouvelle formation.",
                "operationId": "ce924756ce4481d723b589abc508a221",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libeller",
                                    "description",
                                    "durer_formations"
                                ],
                                "properties": {
                                    "libeller": {
                                        "type": "string",
                                        "example": "Nom de la formation"
                                    },
                                    "description": {
                                        "type": "LongText",
                                        "example": "Description de la formation"
                                    },
                                    "durer_formations": {
                                        "type": "string",
                                        "example": "Durée de la formation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formation créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Formations créée avec succès"
                                        },
                                        "data": {
                                            "$ref": "/components/schemas/Formations"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/formations/edit/{formation}": {
            "put": {
                "tags": [
                    "Formations"
                ],
                "summary": "Mettre à jour une formation existante",
                "description": "Permet de mettre à jour une formation existante.",
                "operationId": "da2593fefcbc55f516ef654e606570e7",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "description": "ID de la formation à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libeller",
                                    "description",
                                    "durer_formations"
                                ],
                                "properties": {
                                    "libeller": {
                                        "type": "string",
                                        "example": "Nouveau nom de la formation"
                                    },
                                    "description": {
                                        "type": "Longtext",
                                        "example": "Nouvelle description de la formation"
                                    },
                                    "durer_formations": {
                                        "type": "string",
                                        "example": "Nouvelle durée de la formation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formation modifiée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Formations modifiée avec succès"
                                        },
                                        "data": {
                                            "$ref": "/components/schemas/Formations"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/formations/destroy/{formation}": {
            "delete": {
                "tags": [
                    "Formations"
                ],
                "summary": "Supprimer une formation",
                "description": "Supprime une formation de manière logique ( delete).",
                "operationId": "0cbaa0c1dc62bf9db95eba370015a767",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "description": "ID de la formation à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formation supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "/components/schemas/Formations"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}